+++ src/editor/core/plugins/strong/schema.js
@@ -9,7 +9,7 @@ const schema = {
     marks: {
         strong: {
             sortOrder: 200,
-            parseDom: [{tag: "b"}, {tag: "strong"},
+            parseDOM: [{tag: "b"}, {tag: "strong"},
                 {
                     style: "font-weight", getAttrs: function (value) {
                     return /^(bold(er)?|[5-9]\d{2,})$/.test(value) && null;